🐞 `fix(Makefile): add show-trace flag to deploy target`
Add the `--show-trace` flag to the deploy target in the Makefile to
help diagnose build issues.

📖 `docs(flake.nix): update home-manager URL to release-24.11`

Update the home-manager URL in flake.nix to point to the release-24.11
branch for stability and bug fixes.

🌍 `website(macos.nix): add colima to homebrew packages`

Add the `colima` package to the list of Homebrew packages in the macOS
configuration to support containerized development.

🎨 `style(shared/settings.nix): remove Nixpkgs release check`

Remove the `home.enableNixpkgsReleaseCheck` setting to suppress the
warning about mismatched Home Manager and Nixpkgs versions.